RavenDB-19664 support smuggler/import-dir
