+++ src/Raven.Server/Documents/Handlers/Processors/Smuggler/SmugglerHandlerProcessorForImportDir.cs
@@ -1,0 +1,16 @@
+﻿using JetBrains.Annotations;
+using Raven.Server.ServerWide.Context;
+using Raven.Server.Smuggler.Documents.Handlers;
+
+namespace Raven.Server.Documents.Handlers.Processors.Smuggler;
+
+internal class SmugglerHandlerProcessorForImportDir : AbstractSmugglerHandlerProcessorForImportDir<SmugglerHandler, DocumentsOperationContext>
+{
+    public SmugglerHandlerProcessorForImportDir([NotNull] SmugglerHandler requestHandler) : base(requestHandler)
+    {
+    }
+
+    protected override ImportDelegate DoImport => RequestHandler.DoImportInternalAsync;
+
+    protected override long GetOperationId() => RequestHandler.Database.Operations.GetNextOperationId();
+}
